+++ app/Infrastructure/Controllers/WalletCryptocurrencies/WalletCryptocurrenciesController.php
@@ -1,23 +1,19 @@
 <?php
 
-namespace App\Infrastructure\Controllers;
+namespace App\Infrastructure\Controllers\WalletCryptocurrencies;
 
-use App\Application\Exceptions\WalletNotFoundException;
-use App\Application\Services\WalletBalanceService;
+use App\Application\Services\WalletCryptocurrenciesService;
 use App\Infrastructure\Controllers\Validators\WalletIdValidator;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
 
-class WalletBalanceController extends BaseController
+class WalletCryptocurrenciesController extends BaseController
 {
-    /**
-     * @throws WalletNotFoundException
-     */
     public function __invoke(
         $walletId,
         WalletIdValidator $walletIdValidator,
-        WalletBalanceService $walletBalanceService
+        WalletCryptocurrenciesService $walletCryptocurrenciesService
     ): JsonResponse {
         if (!$walletIdValidator->validateWalletId($walletId)) {
             return response()->json(
@@ -29,13 +25,8 @@ class WalletBalanceController extends BaseController
             );
         }
 
-        $balance = $walletBalanceService->execute($walletId);
+        $walletCryptocurrencies = $walletCryptocurrenciesService->execute($walletId);
 
-        return response()->json(
-            [
-                'balance_usd' => $balance
-            ],
-            Response::HTTP_OK
-        );
+        return response()->json($walletCryptocurrencies, Response::HTTP_OK);
     }
 }
