+++ tests/app/Infrastructure/Controllers/BuyCoin/BuyCoinControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\app\Infrastructure\Controller;
+namespace Tests\app\Infrastructure\Controllers\BuyCoin;
 
 use App\Domain\Coin;
 use App\Domain\DataSources\CoinDataSource;
