+++ tests/app/Infrastructure/Controllers/WalletBalance/WalletBalanceControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\app\Infrastructure\Controller;
+namespace Tests\app\Infrastructure\Controllers\WalletBalance;
 
 use App\Domain\DataSources\CoinDataSource;
 use App\Domain\Wallet;
