+++ tests/app/Infrastructure/Controllers/WalletCryptocurrencies/WalletCryptocurrenciesControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\app\Infrastructure\Controller;
+namespace Tests\app\Infrastructure\Controllers\WalletCryptocurrencies;
 
 use App\Domain\Coin;
 use App\Domain\Wallet;
